+++ SportHouse/index.html
@@ -21,147 +21,9 @@
         margin: 8px;
       "
     >
-      <!-- <section class="container">
-        <h2>Mens Jerssy</h2>
-        <div class="sub-container">
-          <div class="card">
-            <img src="img/18_21.jpg" alt="" />
-            <h4>Jermany Jerrsy</h4>
-            <p>Price: 5000</p>
-
-            <button class="view-detail">View Detail</button>
-          </div>
-          <div class="card">
-            <img src="img/18_21.jpg" alt="" />
-            <h4>Jermany Jerrsy</h4>
-            <p>Price: 5000</p>
-
-            <button class="view-detail">View Detail</button>
-          </div>
-          <div class="card">
-            <img src="img/18_21.jpg" alt="" />
-            <h4>Jermany Jerrsy</h4>
-            <p>Price: 5000</p>
-
-            <button class="view-detail">View Detail</button>
-          </div>
-          <div class="card">
-            <img src="img/18_21.jpg" alt="" />
-            <h4>Jermany Jerrsy</h4>
-            <p>Price: 5000</p>
-
-            <button class="view-detail">View Detail</button>
-          </div>
-        </div>
-      </section> -->
-
-      
-      <!-- <section class="container">
-        <h2>Mens Jerssy</h2>
-        <div class="sub-container">
-          <div class="card">
-            <img src="img/18_21.jpg" alt="" />
-            <h4>Jermany Jerrsy</h4>
-            <p>Price: 5000</p>
-
-            <button class="view-detail">View Detail</button>
-          </div>
-          <div class="card">
-            <img src="img/18_21.jpg" alt="" />
-            <h4>Jermany Jerrsy</h4>
-            <p>Price: 5000</p>
-
-            <button class="view-detail">View Detail</button>
-          </div>
-          <div class="card">
-            <img src="img/18_21.jpg" alt="" />
-            <h4>Jermany Jerrsy</h4>
-            <p>Price: 5000</p>
-
-            <button class="view-detail">View Detail</button>
-          </div>
-          <div class="card">
-            <img src="img/18_21.jpg" alt="" />
-            <h4>Jermany Jerrsy</h4>
-            <p>Price: 5000</p>
-
-            <button class="view-detail">View Detail</button>
-          </div>
-        </div>
-      </section>
-      <section class="container">
-        <h2>Mens Jerssy</h2>
-        <div class="sub-container">
-          <div class="card">
-            <img src="img/18_21.jpg" alt="" />
-            <h4>Jermany Jerrsy</h4>
-            <p>Price: 5000</p>
-
-            <button class="view-detail">View Detail</button>
-          </div>
-          <div class="card">
-            <img src="img/18_21.jpg" alt="" />
-            <h4>Jermany Jerrsy</h4>
-            <p>Price: 5000</p>
-
-            <button class="view-detail">View Detail</button>
-          </div>
-          <div class="card">
-            <img src="img/18_21.jpg" alt="" />
-            <h4>Jermany Jerrsy</h4>
-            <p>Price: 5000</p>
-
-            <button class="view-detail">View Detail</button>
-          </div>
-          <div class="card">
-            <img src="img/18_21.jpg" alt="" />
-            <h4>Jermany Jerrsy</h4>
-            <p>Price: 5000</p>
-
-            <button class="view-detail">View Detail</button>
-          </div>
-        </div>
-      </section>
-      <section class="container">
-        <h2>Mens Jerssy</h2>
-        <div class="sub-container">
-          <div class="card">
-            <img src="img/18_21.jpg" alt="" />
-            <h4>Jermany Jerrsy</h4>
-            <p>Price: 5000</p>
-
-            <button class="view-detail">View Detail</button>
-          </div>
-          <div class="card">
-            <img src="img/18_21.jpg" alt="" />
-            <h4>Jermany Jerrsy</h4>
-            <p>Price: 5000</p>
-
-            <button class="view-detail">View Detail</button>
-          </div>
-          <div class="card">
-            <img src="img/18_21.jpg" alt="" />
-            <h4>Jermany Jerrsy</h4>
-            <p>Price: 5000</p>
-
-            <button class="view-detail">View Detail</button>
-          </div>
-          <div class="card">
-            <img src="img/18_21.jpg" alt="" />
-            <h4>Jermany Jerrsy</h4>
-            <p>Price: 5000</p>
-
-            <button class="view-detail">View Detail</button>
-          </div>
-        </div>
-      </section> -->
     </main>
    
 <footer class="footer"></footer>
-    <script src="js/nav.js"></script>
-    <script src="js/home .js"></script>
-    <script src="js/list.js"></script>
-    <!-- <script src="js/carousel.js"></script> -->
-    <script src="js/footer.js"></script>
+    <script type="module"   src="js/main.js"></script>
   </body>
 </html>
