+++ SportHouse/js/list.js
@@ -1,0 +1,41 @@
+export const productList = [
+
+    {
+   id: 1,
+   title: "Mens Jersy",
+   name: "Jermany Jersy",
+   src: "img/18_21.jpg",
+   price: " price 500",
+   description:
+     "Lorem ipsum dolor sit amet consectetur adipisicing elit. Incidunt, unde vero. Corporis, vero. Doloremque consequuntur, tenetur magni obcaecati distinctio officiis blanditiis recusandae aspernatur illo ipsam quam ex earum dolorum facere.",
+  },
+  {
+    id: 2,
+   title: "Mens Jersy",
+   name: "Brazil Jersy",
+   src: "img/18_21.jpg",
+   price: " price 500",
+   description:
+     "Lorem ipsum dolor sit amet consectetur adipisicing elit. Incidunt, unde vero. Corporis, vero. Doloremque consequuntur, tenetur magni obcaecati distinctio officiis blanditiis recusandae aspernatur illo ipsam quam ex earum dolorum facere.",
+  },
+  {
+   id: 3,
+   title: "Mens Jersy",
+   name: "Portugal Jersy",
+   src: "img/18_21.jpg",
+   price: " price 500",
+   description:
+     "Lorem ipsum dolor sit amet consectetur adipisicing elit. Incidunt, unde vero. Corporis, vero. Doloremque consequuntur, tenetur magni obcaecati distinctio officiis blanditiis recusandae aspernatur illo ipsam quam ex earum dolorum facere.",
+  },
+  {
+   id: 4,
+   title: "Mens Jersy",
+   name: "hungary Jersy",
+   src: "img/18_21.jpg",
+   price: " price 500",
+   description:
+     "Lorem ipsum dolor sit amet consectetur adipisicing elit. Incidunt, unde vero. Corporis, vero. Doloremque consequuntur, tenetur magni obcaecati distinctio officiis blanditiis recusandae aspernatur illo ipsam quam ex earum dolorum facere.",
+  }
+  
+  ]
+
