+++ SportHouse/js/nav.js
@@ -1,8 +1,8 @@
-const createNav = ()=>{
+export const createNav = ()=>{
     let nav = document.querySelector(".navbar");
     nav.innerHTML = `
     <div class="nav">
-    <a href="/"><img src="img/logo.png" class="logo" alt="" /></a>
+    <a href="/SportHouse/index.html"><img src="img/logo.png" class="logo" alt="" /></a>
     <div class="nav-items">
       <div class="search">
         <input
@@ -18,7 +18,7 @@ const createNav = ()=>{
     </div>
   </div>
   <ul class="links-container">
-    <li class="link-item"><a href="#" class="link">home</a></li>
+    <li class="link-item"><a href="/SportHouse/index.html" class="link">home</a></li>
     <li class="link-item"><a href="#" class="link">women</a></li>
     <li class="link-item"><a href="#" class="link">men</a></li>
     <li class="link-item"><a href="#" class="link">kids</a></li>
