+++ SportHouse/reg.html
@@ -20,7 +20,6 @@
         maxlength="20"
         id="firstName"
         class="form-control"
-        value=""
         placeholder="Enter your first name here"
       />
       </div>
@@ -32,7 +31,6 @@
         name="lastname"
         id="lastname"
         class="form-control"
-        value=""
         placeholder="Enter your last name here"
       />
       </div>
@@ -43,7 +41,6 @@
         name="email"
         id="email"
         class="form-control"
-        value=""
         placeholder="Enter your email here"
       />
       </div>
@@ -55,7 +52,6 @@
         minlength="10"
         maxlength="10"
         class="form-control"
-        value=""
         placeholder="Enter your number here"
       />
       </div>
@@ -65,7 +61,6 @@
         name="password"
         id="password"
         class="form-control"
-        value=""
         placeholder="Enter password"
       />
       </div>
@@ -74,8 +69,7 @@
         type="text"
         name="confirmpassword"
         id="confirmpassword"
-        class="form-control"
-        value=""
+        class="form-control"  
         placeholder=" Confirm Password"
       />
       </div>
