+++ SportHouse/singleproductpage.html
@@ -14,13 +14,13 @@
 
     <main class="single-product-main">
       <section class="product-detail">
-        <div class="single-product-card">
-          <img class="single-product-image" src="img/demo1.jpg" alt="" />
+        <div class="single-product-card" id="single-product-card">
+          <!-- <img class="single-product-image" src="img/demo1.jpg" alt="" /> -->
         </div>
         <div class="single-product-info">
             <h3>Germany Jerssay</h2>
             <h4>Price: 500</h4>
-            <form class="quantity" action="" method="post">
+            <form class="cart-form" action="" method="post">
               <button class="decrement-quantity">-</button>
               <input class="product-input" type="text" id="product-input">
               <button class="increment-quantity">+</button>
@@ -34,9 +34,8 @@
       </section>
                 
     </main>
-    <footer class="footer"></footer>
-    <script src="js/nav.js"></script>
-    <script src="js/footer.js"></script>
-    <script src="js/productDetail.js"></script>
+<footer class="footer"></footer>
+
+    <script type="module" src="js/productDetail.js"></script>
   </body>
 </html>
